Refactoring Repository Layer
