+++ app/src/main/java/com/unsa/suppliers/data/repositories/SupplierRepository.kt
@@ -1,35 +1,12 @@
-package com.unsa.suppliers.data
+package com.unsa.suppliers.data.repositories
 
-import com.unsa.suppliers.data.dtos.auth.LoginRequest
-import com.unsa.suppliers.data.dtos.auth.LoginResponse
-import com.unsa.suppliers.data.dtos.auth.UserRequest
-import com.unsa.suppliers.data.dtos.auth.UserResponse
-import com.unsa.suppliers.data.dtos.main.categories.CategoryRequest
-import com.unsa.suppliers.data.dtos.main.categories.CategoryResponse
-import com.unsa.suppliers.data.dtos.main.countries.CountryRequest
-import com.unsa.suppliers.data.dtos.main.countries.CountryResponse
-import com.unsa.suppliers.data.dtos.main.suppliers.SupplierRequest
-import com.unsa.suppliers.data.dtos.main.suppliers.SupplierResponse
+import com.unsa.suppliers.data.dtos.main.suppliers.*
 import com.unsa.suppliers.data.network.services.SupplierService
-import com.unsa.suppliers.data.network.services.AuthService
-import com.unsa.suppliers.data.network.services.CategoryService
-import com.unsa.suppliers.data.network.services.CountryService
 import javax.inject.Inject
 
 class SupplierRepository @Inject constructor (
-    private val authService: AuthService,
     private val supplierService: SupplierService,
-    private val categoryService: CategoryService,
-    private val countryService: CountryService
 ) {
-    // Login Actions
-    suspend fun attemptLogin(loginRequest: LoginRequest): LoginResponse? {
-        return authService.login(loginRequest)
-    }
-    suspend fun attemptRegister(userRequest: UserRequest): UserResponse? {
-        return authService.register(userRequest)
-    }
-    // Suppliers Actions
     suspend fun attemptGetSuppliers(): MutableList<SupplierResponse>? {
         return supplierService.getSuppliers()
     }
@@ -47,43 +24,5 @@ class SupplierRepository @Inject constructor (
     }
     suspend fun attemptDisableSupplier(id: Int) {
         supplierService.disableSupplier(id)
-    }
-    // Categories Actions
-    suspend fun attemptGetCategories(): MutableList<CategoryResponse>? {
-        return categoryService.getAllCategories()
-    }
-    suspend fun attemptGetCategoryById(id: Int): CategoryResponse? {
-        return categoryService.getCategoryById(id)
-    }
-    suspend fun attemptCreateCategory(categoryRequest: CategoryRequest): CategoryResponse? {
-        return categoryService.createCategory(categoryRequest)
-    }
-    suspend fun attemptDeleteCategory(id: Int) {
-        categoryService.deleteCategory(id)
-    }
-    suspend fun attemptEnableCategory(id: Int) {
-        categoryService.enableCategory(id)
-    }
-    suspend fun attemptDisableCategory(id: Int) {
-        categoryService.disableCategory(id)
-    }
-    // Countries Actions
-    suspend fun attemptGetCountries(): MutableList<CountryResponse>? {
-        return countryService.getAllCountries()
-    }
-    suspend fun attemptGetCountryById(id: Int): CountryResponse? {
-        return countryService.getCountryById(id)
-    }
-    suspend fun attemptCreateCountry(countryRequest: CountryRequest): CountryResponse? {
-        return countryService.createCountry(countryRequest)
-    }
-    suspend fun attemptDeleteCountry(id: Int) {
-        countryService.deleteCountry(id)
-    }
-    suspend fun attemptEnableCountry(id: Int) {
-        countryService.enableCountry(id)
-    }
-    suspend fun attemptDisableCountry(id: Int) {
-        countryService.disableCountry(id)
     }
 }
